+++ Assets/OxGFrame/AssetLoader/ThirdParty/YooAsset/Editor/AssetBundleBuilder/BuildPipeline/ScriptableBuildPipeline/ScriptableBuildPipeline.cs
@@ -36,6 +36,7 @@ namespace YooAsset.Editor
                     new TaskCreateReport_SBP(),
                     new TaskCreatePackage_SBP(),
                     new TaskCopyBuildinFiles_SBP(),
+                    new TaskCreateCatalog_SBP()
                 };
             return pipeline;
         }