+++ Assets/OxGFrame/AssetLoader/ThirdParty/YooAsset/Runtime/FileSystem/DefaultBuildinFileSystem/DefaultBuildinFileSystemBuild.cs
@@ -15,15 +15,6 @@ namespace YooAsset
         /// </summary>
         public void OnPreprocessBuild(UnityEditor.Build.Reporting.BuildReport report)
         {
-            ExportBuildinCatalogFile();
-        }
-
-        /// <summary>
-        /// 输出包裹的内置资源目录文件
-        /// </summary>
-        /// <exception cref="System.Exception"></exception>
-        public static void ExportBuildinCatalogFile()
-        {
             YooLogger.Log("Begin to create catalog file !");
 
             string rootPath = YooAssetSettingsData.GetYooDefaultBuildinRoot();
@@ -93,6 +84,7 @@ namespace YooAsset
 
             // 创建内置清单实例
             var buildinFileCatalog = new DefaultBuildinFileCatalog();
+            buildinFileCatalog.FileVersion = CatalogDefine.FileVersion;
             buildinFileCatalog.PackageName = packageName;
             buildinFileCatalog.PackageVersion = packageVersion;
 
@@ -106,7 +98,8 @@ namespace YooAsset
                 $"{packageName}_{packageVersion}.hash",
                 $"{packageName}_{packageVersion}.json",
                 $"{packageName}_{packageVersion}.report",
-                DefaultBuildinFileSystemDefine.BuildinCatalogFileName
+                DefaultBuildinFileSystemDefine.BuildinCatalogJsonFileName,
+                DefaultBuildinFileSystemDefine.BuildinCatalogBinaryFileName
             };
 
             // 记录所有内置资源文件
@@ -123,7 +116,9 @@ namespace YooAsset
                 string fileName = fileInfo.Name;
                 if (fileMapping.TryGetValue(fileName, out string bundleGUID))
                 {
-                    var wrapper = new DefaultBuildinFileCatalog.FileWrapper(bundleGUID, fileName);
+                    var wrapper = new DefaultBuildinFileCatalog.FileWrapper();
+                    wrapper.BundleGUID = bundleGUID;
+                    wrapper.FileName = fileName;
                     buildinFileCatalog.Wrappers.Add(wrapper);
                 }
                 else
@@ -132,16 +127,20 @@ namespace YooAsset
                 }
             }
 
-            // 创建输出目录
-            string saveFilePath = $"{pacakgeDirectory}/{DefaultBuildinFileSystemDefine.BuildinCatalogFileName}";
-            if (File.Exists(saveFilePath))
-                File.Delete(saveFilePath);
+            // 创建输出文件
+            string jsonFilePath = $"{pacakgeDirectory}/{DefaultBuildinFileSystemDefine.BuildinCatalogJsonFileName}";
+            if (File.Exists(jsonFilePath))
+                File.Delete(jsonFilePath);
+            CatalogTools.SerializeToJson(jsonFilePath, buildinFileCatalog);
 
             // 创建输出文件
-            File.WriteAllText(saveFilePath, JsonUtility.ToJson(buildinFileCatalog, false));
-            UnityEditor.AssetDatabase.Refresh();
+            string binaryFilePath = $"{pacakgeDirectory}/{DefaultBuildinFileSystemDefine.BuildinCatalogBinaryFileName}";
+            if (File.Exists(binaryFilePath))
+                File.Delete(binaryFilePath);
+            CatalogTools.SerializeToBinary(binaryFilePath, buildinFileCatalog);
 
-            Debug.Log($"Succeed to save buildin file catalog : {saveFilePath}");
+            UnityEditor.AssetDatabase.Refresh();
+            Debug.Log($"Succeed to save buildin file catalog : {binaryFilePath}");
             return true;
         }
     }