+++ Assets/OxGFrame/AssetLoader/ThirdParty/YooAsset/Runtime/FileSystem/DefaultBuildinFileSystem/DefaultBuildinFileSystemDefine.cs
@@ -4,8 +4,13 @@ namespace YooAsset
     internal class DefaultBuildinFileSystemDefine
     {
         /// <summary>
-        /// 内置清单文件名称
+        /// 内置清单JSON文件名称
         /// </summary>
-        public const string BuildinCatalogFileName = "BuildinCatalog.json";
+        public const string BuildinCatalogJsonFileName = "BuildinCatalog.json";
+
+        /// <summary>
+        /// 内置清单二进制文件名称
+        /// </summary>
+        public const string BuildinCatalogBinaryFileName = "BuildinCatalog.bytes";
     }
 }